+++ CHANGES.md
@@ -1,4 +1,6 @@
 [unreleased]
+
+#### 1.13.0 / 2023-07-10
 * update version check
 * simplify plugin card notice